master2/METHODOLOGIE_TESTS/4_TP/methodologie-du-test/requirements.txt
ruff
pytest
python-dotenv
pandas
